\name{dbNow}
\alias{dbNow}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get the actual system time
}
\description{\code{dbNow} retrieves the acrtual system time.}
\usage{dbNow()}
\details{Returns the actual system time in the format yyyy-mm-dd hh:mm:ss.}
\value{A string.}
\author{K. Gerald van den Boogaart, S. Matos Camacho}
\examples{
dbNow()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{utilities}
